% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4DNavbarMenu.R
\name{bs4DNavbarMenu}
\alias{bs4DNavbarMenu}
\title{Create Navbar Menu (in bs4Dash)}
\usage{
bs4DNavbarMenu(..., id = NULL)
}
\arguments{
\item{...}{inherited arguments}

\item{id}{id for menu}
}
\value{
navbar menu (HTML)
}
\description{
This is a standard navbar menu in bs4Dash covered in this
    GitHub comment/issue:
    https://github.com/RinteRface/bs4Dash/issues/108#issuecomment-772101031
}
\examples{
require(bs4Dash)
  ui <- dashboardPage(
    header = dashboardHeader(
    bs4DNavbarMenu(
        bs4DNavbarTab(tabName = "Tab1", "Tab 1"),
        bs4DNavbarTab(tabName = "Tab2", "Tab 2")
        )
       ),
   body = dashboardBody(
    tabItems(
      tabItem(
        tabName = "Tab1",
                  "Tab 1"
                  ),
         tabItem(
         tabName = "Tab2",
                  "Tab 2"
             )
         )
     ),
 sidebar = dashboardSidebar(
      disable = TRUE)
  )
 server <- function(input, output, session) {

 }
# ()shinyApp(ui = ui, server = server)
}
