% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs4DNavbarTab.R
\name{bs4DNavbarTab}
\alias{bs4DNavbarTab}
\title{Create Navbar Tab (in bs4Dash)}
\usage{
bs4DNavbarTab(tabName, ..., icon = NULL)
}
\arguments{
\item{tabName}{name of tab}

\item{...}{ids for tab}

\item{icon}{fontawesome icon}
}
\value{
navbar tabs (HTML)
}
\description{
Create Navbar Tab (in bs4Dash)
}
\examples{
require(bs4Dash)

  ui <- dashboardPage(
    header = dashboardHeader(
    navbarMenu(
        bs4DNavbarTab(tabName = "Tab1", "Tab 1"),
        bs4DNavbarTab(tabName = "Tab2", "Tab 2")
        )
       ),
   body = dashboardBody(
    tabItems(
      tabItem(
        tabName = "Tab1",
                  "Tab 1"
                  ),
         tabItem(
         tabName = "Tab2",
                  "Tab 2"
             )
         )
     ),
 sidebar = dashboardSidebar(
      disable = TRUE)
  )
 server <- function(input, output, session) {

 }
# ()shinyApp(ui = ui, server = server)
}
