% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tooltip_var_input_ui.R
\name{mod_tooltip_var_input_ui}
\alias{mod_tooltip_var_input_ui}
\title{Create UI inputs with tool tips}
\usage{
mod_tooltip_var_input_ui(id)
}
\arguments{
\item{id}{module ID}
}
\value{
UI module function for tooltip app
}
\description{
Create UI inputs with tool tips
}
