% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{mod_var_input_server}
\alias{mod_var_input_server}
\title{var_input Server Functions}
\usage{
mod_var_input_server(id)
}
\arguments{
\item{id}{}
}
\value{
shiny server module
}
\description{
var_input Server Functions
}
