% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narbarMenuApp.R
\name{navbarMenuApp}
\alias{navbarMenuApp}
\title{Create navbarMenu (bs4Dash)}
\usage{
navbarMenuApp()
}
\value{
shiny app (with navbarMenu)
}
\description{
Create navbarMenu (bs4Dash)
}
\examples{
require(JSbs4dash)
# navbarMenuApp()
}
